feat: Expose interactive command
